@@ -1,5 +1,4 @@
 var fs = require('fs');
-var url = require('url');
 var path = require('path');
 var http = require('http');
 var argv = require('minimist')(process.argv.slice(2));
@@ -10,7 +9,7 @@ function listDirs(root) {
   var files = fs.readdirSync(root);
   var dirs = [];
 
-  for (var i=0, l=files.length; i<l; i++) {
+  for (var i = 0, l = files.length; i < l; i++) {
     var file = files[i];
     if (file[0] !== '.') {
       var stat = fs.statSync(path.join(root, file));
@@ -105,7 +104,7 @@ server = http.createServer(function (req, res) {
   if (/\/$/.test(url)) {
     url += 'index.html';
   }
-  
+
   // Format request /get -> /get/index.html
   var parts = url.split('/');
   if (dirs.indexOf(parts[parts.length - 1]) > -1) {
@@ -134,4 +133,8 @@ server = http.createServer(function (req, res) {
   }
 });
 
-server.listen(argv.p || 3000);
+const PORT = argv.p || 3000;
+
+server.listen(PORT);
+
+console.log("Examples running on " + PORT);
